Work in progress: semester budget setup UI and viewmodels

+++ app/src/main/java/com/example/pocketplan/ui/components/PocketPlanComponents.kt
@@ -601,7 +601,7 @@ fun CategoryStatusRow(
                 color = TextPrimary
             )
             Text(
-                text = CurrencyUtils.formatToUGX(category.allocatedAmount),
+                text = CurrencyUtils.formatToUGX(category.allocatedAmount.toDouble()),
                 style = MaterialTheme.typography.bodySmall,
                 color = TextSecondary
             )
